EPA-analysis/requirements.txt
requests>=2.25.0
python-dotenv>=0.19.0
tqdm>=4.64.0
argparse>=1.4.0
openai>=0.28.0
numpy>=1.20.0
supabase>=2.0.0
scikit-learn>=1.0.0
pandas>=1.3.0